+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Added library and header files to install
+- Improved output when a key is set multiple times
 
 ## 0.1.0 (2020-12-08)